Move create-client-dialog.component and associated API call to AC
