+++ libs/common/src/admin-console/abstractions/provider/provider-api.service.abstraction.ts
@@ -2,12 +2,12 @@ import { AddableOrganizationResponse } from "@bitwarden/common/admin-console/mod
 import { ProviderOrganizationOrganizationDetailsResponse } from "@bitwarden/common/admin-console/models/response/provider/provider-organization.response";
 import { ListResponse } from "@bitwarden/common/models/response/list.response";
 
+import { CreateProviderOrganizationRequest } from "../../models/request/create-provider-organization.request";
 import { ProviderSetupRequest } from "../../models/request/provider/provider-setup.request";
 import { ProviderUpdateRequest } from "../../models/request/provider/provider-update.request";
 import { ProviderVerifyRecoverDeleteRequest } from "../../models/request/provider/provider-verify-recover-delete.request";
 import { UpdateProviderOrganizationRequest } from "../../models/request/update-provider-organization.request";
 import { ProviderResponse } from "../../models/response/provider/provider.response";
-
 
 export abstract class ProviderApiServiceAbstraction {
   abstract postProviderSetup(id: string, request: ProviderSetupRequest): Promise<ProviderResponse>;
@@ -37,4 +37,9 @@ export abstract class ProviderApiServiceAbstraction {
     organizationId: string,
     request: UpdateProviderOrganizationRequest,
   ): Promise<any>;
+
+  abstract createProviderOrganization(
+    providerId: string,
+    request: CreateProviderOrganizationRequest,
+  ): Promise<void>;
 }
