+++ libs/common/src/admin-console/models/request/create-provider-organization.request.ts
@@ -1,9 +1,10 @@
 // FIXME: Update this file to be type safe and remove this and next line
 // @ts-strict-ignore
-import { OrganizationKeysRequest } from "../../../admin-console/models/request/organization-keys.request";
 import { PlanType } from "../../../billing/enums";
 
-export class CreateClientOrganizationRequest {
+import { OrganizationKeysRequest } from "./organization-keys.request";
+
+export class CreateProviderOrganizationRequest {
   name: string;
   ownerEmail: string;
   planType: PlanType;
