+++ libs/common/src/admin-console/services/provider/provider-api.service.ts
@@ -4,12 +4,12 @@ import { ListResponse } from "@bitwarden/common/models/response/list.response";
 
 import { ApiService } from "../../../abstractions/api.service";
 import { ProviderApiServiceAbstraction } from "../../abstractions/provider/provider-api.service.abstraction";
+import { CreateProviderOrganizationRequest } from "../../models/request/create-provider-organization.request";
 import { ProviderSetupRequest } from "../../models/request/provider/provider-setup.request";
 import { ProviderUpdateRequest } from "../../models/request/provider/provider-update.request";
 import { ProviderVerifyRecoverDeleteRequest } from "../../models/request/provider/provider-verify-recover-delete.request";
 import { UpdateProviderOrganizationRequest } from "../../models/request/update-provider-organization.request";
 import { ProviderResponse } from "../../models/response/provider/provider.response";
-
 
 export class ProviderApiService implements ProviderApiServiceAbstraction {
   constructor(private apiService: ApiService) {}
@@ -102,6 +102,19 @@ export class ProviderApiService implements ProviderApiServiceAbstraction {
     return await this.apiService.send(
       "PUT",
       "/providers/" + providerId + "/clients/" + organizationId,
+      request,
+      true,
+      false,
+    );
+  }
+
+  createProviderOrganization(
+    providerId: string,
+    request: CreateProviderOrganizationRequest,
+  ): Promise<void> {
+    return this.apiService.send(
+      "POST",
+      "/providers/" + providerId + "/clients",
       request,
       true,
       false,
